# backupccl: parallelize loading of manifests from External Storage