@adityamaru 2022-09-02
Comparing two traces of a `SHOW BACKUP` on a backup chain that contains s a 1.80TB (replicated) backup consisting of 50 incremental backup layers taken every 15 minutes. 8000 warehouses were imported and then a workload of 1000 warehouses was run against the cluster while the incremental backups were being taken.

Old trace:

This is a summary of where we are spending time in this operation:

```
     0.000ms      0.000ms    === operation:traced statement _verbose:1 node:1 client:[::1]:59791 user:root
     0.000ms      0.000ms    [flow: {count: 1, duration 3m6s}]
     0.000ms      0.000ms    [batch flow coordinator: {count: 1, duration 3m6s}]
     0.000ms      0.000ms    [backupccl.showBackupPlanHook: {count: 1, duration 3m6s}]
     0.000ms      0.000ms    [SHOW BACKUP: {count: 1, duration 3m6s}]
     0.000ms      0.000ms    [backupdest.FindPriorBackups: {count: 3, duration 2m46s}]
     0.000ms      0.000ms    [gcs.List: {count: 3, duration 2m46s}]
     0.000ms      0.000ms    [backupdest.ResolveBackupManifests: {count: 1, duration 1m45s}]
     0.000ms      0.000ms    [backupdest.ResolveIncrementalsBackupLocation: {count: 1, duration 1m20s}]
     0.000ms      0.000ms    [backupdest.backupsFromLocation: {count: 2, duration 1m20s}]
     0.000ms      0.000ms    [backupinfo.ReadBackupManifest: {count: 92, duration 19s}]
     0.000ms      0.000ms    [gcs.ReadFileAt: {count: 184, duration 18.7s}]
     0.000ms      0.000ms    [backupdest.LayerLoop: {count: 1, duration 18.5s}]
     0.000ms      0.000ms    [backupccl.showBackup: {count: 1, duration 507ms}]
     0.000ms      0.000ms    [backupinfo.ReadBackupManifestFromStore: {count: 1, duration 470ms}]
     0.000ms      0.000ms    [backupccl.backupShowerDefault.fn: {count: 1, duration 386ms}]
     0.000ms      0.000ms    [backupinfo.readManifest: {count: 92, duration 258ms}]
     0.000ms      0.000ms    [backupinfo.GetLocalityInfo: {count: 92, duration 388µs}]
     0.000ms      0.000ms    [optimizer: {count: 1, duration 286µs}]
     0.000ms      0.000ms    [commit sql txn: {count: 1, duration 10µs}]
```

As you can see the majority of the time is spent in `backupdest.ResolveBackupManifests` and `backupdest.ResolveIncrementalsBackupLocation` . So if we drill down into those we see:

```
 1.870ms      1.075ms                    === operation:backupdest.ResolveIncrementalsBackupLocation _verbose:1 node:1 client:[::1]:59791 user:root
     1.870ms      0.000ms                    [backupdest.backupsFromLocation: {count: 2, duration 1m20s}]
     1.870ms      0.000ms                    [backupdest.FindPriorBackups: {count: 2, duration 1m20s}]
     1.870ms      0.000ms                    [gcs.List: {count: 2, duration 1m20s}]
```

and

```
80450.060ms    268.367ms                    === operation:backupdest.ResolveBackupManifests _verbose:1 node:1 client:[::1]:59791 user:root
 80450.060ms      0.000ms                    [backupdest.FindPriorBackups: {count: 1, duration 1m26s}]
 80450.060ms      0.000ms                    [gcs.List: {count: 1, duration 1m26s}]
 80450.060ms      0.000ms                    [backupinfo.ReadBackupManifest: {count: 92, duration 19s}]
 80450.060ms      0.000ms                    [gcs.ReadFileAt: {count: 184, duration 18.7s}]
 80450.060ms      0.000ms                    [backupdest.LayerLoop: {count: 1, duration 18.5s}]
 80450.060ms      0.000ms                    [backupinfo.ReadBackupManifestFromStore: {count: 1, duration 470ms}]
 80450.060ms      0.000ms                    [backupinfo.readManifest: {count: 92, duration 258ms}]
 80450.060ms      0.000ms                    [backupinfo.GetLocalityInfo: {count: 92, duration 388µs}]
```

The majority of the time in both methods is spent in the call to `gcs.List` which requires investigation as to whether we can optimize something here, but this diff is concerned with the tracing span called `LayerLoop` which wraps the logic that reaches out to External Storage for every incremental layer and reads the manifest.

New trace:

In the new trace where we have changed the `LayerLoop` logic to load the manifests concurrently we see it take 649ms instead of 18s:

```
76883.700ms    360.265ms                    === operation:backupdest.ResolveBackupManifests _verbose:1 node:1 client:[::1]:60360 user:root
 76883.700ms      0.000ms                    [backupdest.FindPriorBackups: {count: 1, duration 1m25s}]
 76883.700ms      0.000ms                    [gcs.List: {count: 1, duration 1m25s}]
 76883.700ms      0.000ms                    [backupinfo.ReadBackupManifestFromStore: {count: 92, duration 50.6s}]
 76883.700ms      0.000ms                    [backupinfo.ReadBackupManifest: {count: 92, duration 50.6s}]
 76883.700ms      0.000ms                    [gcs.ReadFileAt: {count: 184, duration 50.5s}]
 76883.700ms      0.000ms                    [backupdest.LayerLoop: {count: 1, duration 649ms}]
 76883.700ms      0.000ms                    [backupinfo.FetchPreviousBackups: {count: 1, duration 648ms}]
 76883.700ms      0.000ms                    [backupinfo.readManifest: {count: 92, duration 122ms}]
 76883.700ms      0.000ms                    [backupinfo.GetLocalityInfo: {count: 92, duration 39µs}]
 76883.761ms      0.061ms                        === operation:backupinfo.ReadBackupManifestFromStore _ve
```

The total time goes down for this `SHOW` goes down by 18s. This is far from the dominating cost of the list calls, but its still worth doing.