@adityamaru 2022-09-06
Note, `List` taking so long is a result of the fixture I ran it on being old, and not having some of the delimiter optimizations we checked in to solve this exact problem. Stay tuned for the updated traces once I've regenerated the fixtures.