@joshimhoff 2022-09-06
Yay to this PR!

@adityamaru, can we check what this PR does to the ~2s `SHOW BACKUP` runtime that @dasrirez found originally? Aditya, I'm sure your testing here is a better stress test of `SHOW BACKUP` in general, but OTOH the ~2s  `SHOW BACKUP` runtime being too high is the thing that could actually delay the launch of `backupper` deprecation, so good to know where we stand there.

Darius tells me he setup a local repro that queries some GCS buckets he set up already. Darius, can you send Aditya the SQL command you run on your local cluster? You mentioned that Aditya already has access to credentials for the bucket you setup.