@adityamaru 2022-09-06
> Aditya, I'm sure the testing you've done here is a better stress test of SHOW BACKUP in general, but OTOH the ~2s SHOW BACKUP runtime being too high is the thing that could actually delay the launch of backupper deprecation, so good to know where we stand there.

Definitely makes sense to run on what we actually want to improve. @dasrirez if you want to take this patch for a spin and grab a trace using `EXPLAIN ANALYZE (DEBUG) <SHOW BACKUP query>` that'd be helpful. Alternatively, you could DM me what you've been running so that we're testing on the same corpus. I've been using a pre-generated fixture we run our roachtests against that has 1.7Tb replicated data backed up over a chain of 50 incremental backups.