@dasrirez 2022-09-07
This is wonderful! I ran the same simple test with 1 full + 24 incremental backups taken from a fresh cluster and the difference is absurd.

- Without patch
   - 1 incremental: ~0.88s
   - 24 incrementals: ~4.5s
- With patch
   - 1 incremental: ~1s
   - __24 incrementals__: ~1.4s

Note that I issued the queries from a local cluster (avg of 8 trials) and the bucket lookup latency is much higher than when these queries are issued on the cloud cluster. On the cloud cluster, a backup with a single incremental takes 0.3s to query on average, which is why the backup with 24 incrementals took 2s without the patch compared to the 4.5s I got from my local cluster.